Add a class Item ItemOrganizer
